\name{BertomeuMaMarinovic-package}
\alias{BertomeuMaMarinovic-package}
\alias{BertomeuMaMarinovic}
\docType{package}
\title{
\packageTitle{BertomeuMaMarinovic}
}
\description{
\packageDescription{BertomeuMaMarinovic}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BertomeuMaMarinovic}
\packageIndices{BertomeuMaMarinovic}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{BertomeuMaMarinovic}

Maintainer: \packageMaintainer{BertomeuMaMarinovic}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
